BLE: Move GattServer implementation from generic to cordio
